requirements.txt
pandas 
numpy 
matplotlib 
seaborn
scikit-learn
xgboost
streamlit
sqlalchemy
python-dotenv
# -e .